car_accident_detector/requirements.txt
smbus2>=0.4.1
pyserial>=3.5
pynmea2>=1.15.0
numpy>=1.21.0
tensorflow>=2.8.0
tensorflow-lite>=2.8.0
scikit-learn>=1.0.0
pandas>=1.3.0
RPi.GPIO>=0.7.0